Enforce leading zero for numbers SQL Server only returns ".00" if a value is zero

+++ lib/mshoplib/src/MShop/Price/Item/Base.php
@@ -45,7 +45,7 @@ abstract class Base
 	{
 		parent::__construct( $prefix, $values );
 
-		$this->precision = ( isset( $values['.precision'] ) ? $values['.precision'] : 2 );
+		$this->precision = (int) ( isset( $values['.precision'] ) ? $values['.precision'] : 2 );
 		$this->initListItems( $listItems, $refItems );
 		$this->initPropertyItems( $propItems );
 	}
